+++ src/mako/utility/exceptions/StatisticsException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\utility\exceptions;
+
+use RuntimeException;
+
+/**
+ * Statistics exception.
+ */
+class StatisticsException extends RuntimeException
+{
+
+}